+++ atoma-vllm/src/worker.rs
@@ -185,8 +185,8 @@ where
     #[instrument(skip_all)]
     pub fn cache_swap(
         &mut self,
-        blocks_to_swap_in: &HashMap<i64, i64>,
-        blocks_to_swap_out: &HashMap<i64, i64>,
+        blocks_to_swap_in: &HashMap<u32, u32>,
+        blocks_to_swap_out: &HashMap<u32, u32>,
         blocks_to_copy: Option<Tensor>,
     ) -> Result<(), ModelWorkerError> {
         if blocks_to_swap_in.len() > 0 {
@@ -559,7 +559,7 @@ impl CacheEngine {
     #[instrument(skip_all)]
     pub fn swap_in(
         &mut self,
-        blocks_to_swap_in: &HashMap<i64, i64>,
+        blocks_to_swap_in: &HashMap<u32, u32>,
     ) -> Result<(), CacheEngineError> {
         let _enter = self.span.enter();
         for i in 0..self.num_layers {
@@ -576,7 +576,7 @@ impl CacheEngine {
     #[instrument(skip_all)]
     pub fn swap_out(
         &mut self,
-        blocks_to_swap_out: &HashMap<i64, i64>,
+        blocks_to_swap_out: &HashMap<u32, u32>,
     ) -> Result<(), CacheEngineError> {
         let _enter = self.span.enter();
         for i in 0..self.num_layers {
